Enable distro agnostic build of System.Security.Cryptography.Native (#13885)
* Enable distro agnostic build of System.Security.Cryptography.Native

If the FEATURE_DISTRO_AGNOSTIC_SSL is enabled, the System.Security.Cryptography.Native
doesn't link to libssl / libcrypto at build time. It uses a shim that opens the
libssl shared library with versioned so name specific for the platform and gets all needed
API function pointers using dlsym and stores them in global variables.
The calls to those APIs are then redirected through these pointers.

I have verified that our code doesn't access any data members in structures with layout
that can vary based on compilation options of the OpenSSL and that can have variable
offset. The only direct access of this kind that we were using, the X509::sha1_hash,
was replaced by call to recompute the hash, based on @bartonjs suggestion.

There are couple of functions that we use and that don't have to exist in the libssl when
it is compiled with particular options. These were handled at the build time before, now
I have also added runtime handling of this case.

Finally, I have added verification of the System.Security.Cryptography.Native.OpenSsl.so
to make sure it doesn't have any undefined symbols. This could happen if someone adds
usage of a new OpenSSL API and doesn't add its name to the opensslshim.h. The build fails
in this case and lists all the undefined symbols.

Conflicts:
	src/Native/Unix/System.Security.Cryptography.Native/openssl.cpp
	src/Native/Unix/System.Security.Cryptography.Native/pal_dsa.cpp
	src/Native/Unix/System.Security.Cryptography.Native/pal_dsa.h
	src/Native/Unix/System.Security.Cryptography.Native/pal_evp_pkey_dsa.h