Port Detect HTTP2 and multiplexing support for CURL at runtime
This change allows portable version of dotnet core to run on versions
of CURL with and without HTTP2 and multiplexing support and still be
able to use those features if running on CURL that supports them.
The compile time checks are now made at runtime and detect the feature
support based on comparison of the current version of CURL to the
well known version number that added that feature support. The version
numbers are listed in docs/libcurl/symbols-in-versions in libcurl source
tree.